# internal/docs/swagger.yaml
basePath: /
definitions:
  entity.Task:
    properties:
      approvers:
        items:
          $ref: '#/definitions/entity.User'
        type: array
      author:
        $ref: '#/definitions/entity.User'
      body:
        type: string
      descr:
        type: string
      finished:
        type: boolean
      id:
        type: integer
    type: object
  entity.User:
    properties:
      email:
        type: string
      id:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: alex@mail.ru
    name: API Support
    url: https://gitlab.com/g6834/team3/task
  description: This is a team3 task.service API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Team3.Task.service API
  version: "1.0"
paths:
  /ping:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: accessailability task.service
  /task:
    get:
      description: Get one tasks
      parameters:
      - description: Task ID
        format: string
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            type: string
      summary: get one task
    post:
      consumes:
      - application/json
      description: add (create) new task
      parameters:
      - description: New Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/entity.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            type: string
      summary: create task
  /tasks:
    get:
      description: Get list of tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/entity.Task'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            type: string
      summary: list of task
swagger: "2.0"